[Organization settings] Bisect General settings (#19790)
* Divide General settings & Policies pages

* Add Enterprise callout

* style improvements

* Apply style suggestions from @corygrunk

* button style changes
