+++ components/dashboard/src/teams/TeamSettings.tsx
@@ -10,7 +10,6 @@ import Alert from "../components/Alert";
 import ConfirmationModal from "../components/ConfirmationModal";
 import { InputWithCopy } from "../components/InputWithCopy";
 import Modal, { ModalBody, ModalFooter, ModalHeader } from "../components/Modal";
-import { CheckboxInputField } from "../components/forms/CheckboxInputField";
 import { InputField } from "../components/forms/InputField";
 import { TextInputField } from "../components/forms/TextInputField";
 import { Heading2, Heading3, Subheading } from "../components/typography/headings";
@@ -28,20 +27,12 @@ import { OrgSettingsPage } from "./OrgSettingsPage";
 import { ErrorCode } from "@gitpod/gitpod-protocol/lib/messaging/error";
 import { Button } from "@podkit/buttons/Button";
 import { useInstallationDefaultWorkspaceImageQuery } from "../data/installation/default-workspace-image-query";
-import { useAllowedWorkspaceClassesMemo } from "../data/workspaces/workspace-classes-query";
 import { ConfigurationSettingsField } from "../repositories/detail/ConfigurationSettingsField";
-import {
-    WorkspaceClassesModifyModal,
-    WorkspaceClassesModifyModalProps,
-    WorkspaceClassesOptions,
-} from "../components/WorkspaceClassesOptions";
-import { useMutation } from "@tanstack/react-query";
-import { useAllowedWorkspaceEditorsMemo } from "../data/ide-options/ide-options-query";
-import { IdeOptions, IdeOptionsModifyModal, IdeOptionsModifyModalProps } from "../components/IdeOptions";
-import { useFeatureFlag } from "../data/featureflag-query";
 import { Select, SelectContent, SelectItem, SelectTrigger, SelectValue } from "@podkit/select/Select";
+import { useDocumentTitle } from "../hooks/use-document-title";
 
 export default function TeamSettingsPage() {
+    useDocumentTitle("Organization Settings - General");
     const user = useCurrentUser();
     const org = useCurrentOrg().data;
     const isOwner = useIsOwner();
@@ -51,7 +42,6 @@ export default function TeamSettingsPage() {
     const [teamName, setTeamName] = useState(org?.name || "");
     const [updated, setUpdated] = useState(false);
     const updateOrg = useUpdateOrgMutation();
-    const orgLevelEditorRestrictionEnabled = useFeatureFlag("org_level_editor_restriction_enabled");
 
     const close = () => setModal(false);
 
@@ -130,8 +120,10 @@ export default function TeamSettingsPage() {
             <OrgSettingsPage>
                 <div className="space-y-4">
                     <div>
-                        <Heading2>Organization Details</Heading2>
-                        <Subheading>Details of your organization within Gitpod.</Subheading>
+                        <Heading2>General</Heading2>
+                        <Subheading>
+                            Set the default role and workspace image, name or delete your organization.
+                        </Subheading>
                     </div>
                     <ConfigurationSettingsField>
                         {updateOrg.isError && (
@@ -197,25 +189,6 @@ export default function TeamSettingsPage() {
                     </ConfigurationSettingsField>
 
                     <ConfigurationSettingsField>
-                        <Heading3>Collaboration and sharing</Heading3>
-
-                        {updateTeamSettings.isError && (
-                            <Alert type="error" closable={true} className="mb-2 max-w-xl rounded-md">
-                                <span>Failed to update organization settings: </span>
-                                <span>{updateTeamSettings.error.message || "unknown error"}</span>
-                            </Alert>
-                        )}
-
-                        <CheckboxInputField
-                            label="Workspace Sharing"
-                            hint="Allow workspaces created within an Organization to share the workspace with any authenticated user."
-                            checked={!settings?.workspaceSharingDisabled}
-                            onChange={(checked) => handleUpdateTeamSettings({ workspaceSharingDisabled: !checked })}
-                            disabled={isLoading || !isOwner}
-                        />
-                    </ConfigurationSettingsField>
-
-                    <ConfigurationSettingsField>
                         <Heading3>Workspace images</Heading3>
                         <Subheading>Choose a default image for all workspaces in the organization.</Subheading>
 
@@ -232,20 +205,6 @@ export default function TeamSettingsPage() {
                             settings={settings}
                             installationDefaultWorkspaceImage={installationDefaultImage}
                             onClose={() => setShowImageEditModal(false)}
-                        />
-                    )}
-
-                    <OrgWorkspaceClassesOptions
-                        isOwner={isOwner}
-                        settings={settings}
-                        handleUpdateTeamSettings={handleUpdateTeamSettings}
-                    />
-
-                    {orgLevelEditorRestrictionEnabled && (
-                        <EditorOptions
-                            isOwner={isOwner}
-                            settings={settings}
-                            handleUpdateTeamSettings={handleUpdateTeamSettings}
                         />
                     )}
 
@@ -447,164 +406,3 @@ function OrgDefaultWorkspaceImageModal(props: OrgDefaultWorkspaceImageModalProps
         </Modal>
     );
 }
-
-interface OrgWorkspaceClassesOptionsProps {
-    isOwner: boolean;
-    settings?: OrganizationSettings;
-    handleUpdateTeamSettings: (
-        newSettings: Partial<OrganizationSettings>,
-        options?: { throwMutateError?: boolean },
-    ) => Promise<void>;
-}
-const OrgWorkspaceClassesOptions = ({
-    isOwner,
-    settings,
-    handleUpdateTeamSettings,
-}: OrgWorkspaceClassesOptionsProps) => {
-    const [showModal, setShowModal] = useState(false);
-    const { data: allowedClassesInOrganization, isLoading: isLoadingClsInOrg } = useAllowedWorkspaceClassesMemo(
-        undefined,
-        {
-            filterOutDisabled: true,
-            ignoreScope: ["configuration"],
-        },
-    );
-    const { data: allowedClassesInInstallation, isLoading: isLoadingClsInInstall } = useAllowedWorkspaceClassesMemo(
-        undefined,
-        {
-            filterOutDisabled: true,
-            ignoreScope: ["organization", "configuration"],
-        },
-    );
-
-    const restrictedWorkspaceClasses = useMemo(() => {
-        const allowedList = settings?.allowedWorkspaceClasses ?? [];
-        if (allowedList.length === 0) {
-            return [];
-        }
-        return allowedClassesInInstallation.filter((cls) => !allowedList.includes(cls.id)).map((cls) => cls.id);
-    }, [settings?.allowedWorkspaceClasses, allowedClassesInInstallation]);
-
-    const updateMutation: WorkspaceClassesModifyModalProps["updateMutation"] = useMutation({
-        mutationFn: async ({ restrictedWorkspaceClasses }) => {
-            let allowedWorkspaceClasses = allowedClassesInInstallation.map((e) => e.id);
-            if (restrictedWorkspaceClasses.length > 0) {
-                allowedWorkspaceClasses = allowedWorkspaceClasses.filter(
-                    (e) => !restrictedWorkspaceClasses.includes(e),
-                );
-            }
-            const allAllowed = allowedClassesInInstallation.every((e) => allowedWorkspaceClasses.includes(e.id));
-            if (allAllowed) {
-                // empty means allow all classes
-                allowedWorkspaceClasses = [];
-            }
-            await handleUpdateTeamSettings({ allowedWorkspaceClasses }, { throwMutateError: true });
-        },
-    });
-
-    return (
-        <ConfigurationSettingsField>
-            <Heading3>Available workspace classes</Heading3>
-            <Subheading>
-                Limit the available workspace classes in your organization. Requires{" "}
-                <span className="font-medium">Owner</span> permissions to change.
-            </Subheading>
-
-            <WorkspaceClassesOptions
-                isLoading={isLoadingClsInOrg}
-                className="mt-4"
-                classes={allowedClassesInOrganization}
-            />
-
-            {isOwner && (
-                <Button className="mt-6" onClick={() => setShowModal(true)}>
-                    Manage Classes
-                </Button>
-            )}
-
-            {showModal && (
-                <WorkspaceClassesModifyModal
-                    isLoading={isLoadingClsInInstall}
-                    showSetDefaultButton={false}
-                    showSwitchTitle={false}
-                    restrictedWorkspaceClasses={restrictedWorkspaceClasses}
-                    allowedClasses={allowedClassesInInstallation}
-                    updateMutation={updateMutation}
-                    onClose={() => setShowModal(false)}
-                />
-            )}
-        </ConfigurationSettingsField>
-    );
-};
-
-interface EditorOptionsProps {
-    settings: OrganizationSettings | undefined;
-    isOwner: boolean;
-    handleUpdateTeamSettings: (
-        newSettings: Partial<OrganizationSettings>,
-        options?: { throwMutateError?: boolean },
-    ) => Promise<void>;
-}
-const EditorOptions = ({ isOwner, settings, handleUpdateTeamSettings }: EditorOptionsProps) => {
-    const [showModal, setShowModal] = useState(false);
-    const { data: installationOptions, isLoading: installationOptionsIsLoading } = useAllowedWorkspaceEditorsMemo(
-        undefined,
-        {
-            filterOutDisabled: true,
-            ignoreScope: ["organization", "configuration"],
-        },
-    );
-    const { data: orgOptions, isLoading: orgOptionsIsLoading } = useAllowedWorkspaceEditorsMemo(undefined, {
-        filterOutDisabled: true,
-        ignoreScope: ["configuration"],
-    });
-
-    const updateMutation: IdeOptionsModifyModalProps["updateMutation"] = useMutation({
-        mutationFn: async ({ restrictedEditors, pinnedEditorVersions }) => {
-            const updatedRestrictedEditors = [...restrictedEditors.keys()];
-            const updatedPinnedEditorVersions = Object.fromEntries(pinnedEditorVersions.entries());
-
-            await handleUpdateTeamSettings(
-                { restrictedEditorNames: updatedRestrictedEditors, pinnedEditorVersions: updatedPinnedEditorVersions },
-                { throwMutateError: true },
-            );
-        },
-    });
-
-    const restrictedEditors = new Set<string>(settings?.restrictedEditorNames || []);
-    const pinnedEditorVersions = new Map<string, string>(Object.entries(settings?.pinnedEditorVersions || {}));
-
-    return (
-        <ConfigurationSettingsField>
-            <Heading3>Available editors</Heading3>
-            <Subheading>
-                Limit the available editors in your organization. Requires <span className="font-medium">Owner</span>{" "}
-                permissions to change.
-            </Subheading>
-
-            <IdeOptions
-                isLoading={orgOptionsIsLoading}
-                className="mt-4"
-                ideOptions={orgOptions}
-                pinnedEditorVersions={pinnedEditorVersions}
-            />
-
-            {isOwner && (
-                <Button className="mt-6" onClick={() => setShowModal(true)}>
-                    Manage Editors
-                </Button>
-            )}
-
-            {showModal && (
-                <IdeOptionsModifyModal
-                    isLoading={installationOptionsIsLoading}
-                    ideOptions={installationOptions}
-                    restrictedEditors={restrictedEditors}
-                    pinnedEditorVersions={pinnedEditorVersions}
-                    updateMutation={updateMutation}
-                    onClose={() => setShowModal(false)}
-                />
-            )}
-        </ConfigurationSettingsField>
-    );
-};
